add damage test, setPos test, don't move when shot, cleanup existing
